+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-From v2.0.1 to v2.1.0.dev0
+From v2.0.1 to v2.1.0.a2
 
 ## Fixes
 
@@ -17,6 +17,11 @@ From v2.0.1 to v2.1.0.dev0
 - In the case single-task dataset, the `RandomSplit` now uses `StratifiedShuffleSplit` in case of classification.
 - `DuplicatesFilter` refactored to`RepeatsFilter`, as it also captures scenarios where triplicates/quadruplicates are found in the dataset. These scenarios are now also covered by the respective UnitTest.
 - The versioning scheme of development snapshots has changed from `devX` to `alphaX`/`betaX`, where `X` is an integer that increments with each release.
+- The following model class have been renamed and moved:
+    - `models.models.QSPRsklearn` > `models.sklearn.SklearnModel`
+    - `deep.models.QSPRDNN` > `extra.gpu.models.dnn.DNNModel`
+    - `extra.models.pcm.ModelPCM` > `extra.models.pcm.PCMModel`
+    - `extra.models.pcm.QSPRsklearnPCM` > `extra.models.pcm.SklearnPCMModel`
 
 ## New Features
 - `ClusterSplit` - splits data based clustering of molecular fingerprints.
@@ -31,8 +36,8 @@ From v2.0.1 to v2.1.0.dev0
 - New module `early_stopping` with classes `EarlyStopping` and `EarlyStoppingMode` has been added. This module allows for more control over early stopping in models that support it.
 - Add new descriptorset `SmilesDesc` to use the smiles strings as a descriptor.
 - Refactoring of the test suite under `qsprpred.data` and improvement of temporary file handling (!114).
-- `PyBoostModel` - QSPRpred wrapper for py-boost models.
-- `Chemprop` - QSPRpred wrapper for Chemprop model.
+- `PyBoostModel` - QSPRpred wrapper for py-boost models. Requires optional `pyboost` dependencies.
+- `ChempropModel` - QSPRpred wrapper for Chemprop models. Requires optional `deep` dependencies.
 
 ## Removed Features
 - `StratifiedPerTarget` is replaced by `PCMSplit`.